refactor: organize select areas layer files
