+++ app/components/Map.tsx
@@ -14,18 +14,16 @@ import { PlusIcon } from "@phosphor-icons/react";
 import { useColorModeValue } from "./ui/color-mode";
 import useMapStore from "@/app/store/mapStore";
 import MapAreaControls from "./MapAreaControls";
-import SelectAreaLayer from "./map/layers/SelectAreaLayer";
 import useContextStore from "@/app/store/contextStore";
-import CustomAreasLayer from "./map/layers/CustomAreasLayer";
 import HighlightedFeaturesLayer from "./map/layers/HighlightedFeaturesLayer";
+import SelectAreaLayer from "./map/layers/select-area-layer";
 
 const MAPBOX_ACCESS_TOKEN = process.env.NEXT_PUBLIC_MAPBOX_ACCESS_TOKEN;
 
 function Map() {
   const mapRef = useRef<MapRef>(null);
   const [mapCenter, setMapCenter] = useState([0, 0]);
-  const { geoJsonFeatures, setMapRef, selectAreaLayer, initializeTerraDraw } =
-    useMapStore();
+  const { geoJsonFeatures, setMapRef, initializeTerraDraw } = useMapStore();
   const { context } = useContextStore();
   const areas = context.filter((c) => c.contextType === "area");
 
@@ -115,14 +113,7 @@ function Map() {
           areas={areas}
         />
 
-        {selectAreaLayer && selectAreaLayer !== "Custom" && (
-          <SelectAreaLayer
-            key={selectAreaLayer}
-            layerId={selectAreaLayer}
-            beforeId={undefined}
-          />
-        )}
-        {selectAreaLayer === "Custom" && <CustomAreasLayer />}
+        <SelectAreaLayer />
         <MapAreaControls />
 
         <AttributionControl
